+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 - Add rs_driver_viewer, a small tool to show point cloud
 - Add save_by_rows argument
 - Add multi-cast support
-
+- Add points transformation function
 ### Changed
 
 - Update some decoding part for LiDARs